static_cast to remove warnings; JIRA: PGAP-9240
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102893 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/asn_cache/asn_cache_test.cpp
@@ -436,7 +436,7 @@ int CAsnCacheTestApplication::Run(void)
                  << count / e << " seqs/sec ("
                  << count_failed << " failed to retrieve)");
     }
-    if(args["dump-proteins"]) { return count_failed; }
+    if(args["dump-proteins"]) { return static_cast<int>(count_failed); }
     if(serialize) {
         if(getIdOnly){
             ITERATE(vector< CDataLoader::TIds >, it, id_sets){
